+++ src/suou/sqlalchemy/__init__.py
@@ -158,7 +158,7 @@ from .asyncio import SQLAlchemy, async_query, SessionWrapper, AsyncSelectPaginat
 from .orm import (
     id_column, snowflake_column, match_column, match_constraint, bool_column, declarative_base, parent_children,
     author_pair, age_pair, bound_fk, unbound_fk, want_column, a_relationship, BitSelector, secret_column, username_column,
-    ascii_column
+    ascii_column, i4_column
 )
 
 try:
@@ -173,7 +173,7 @@ __all__ = (
     'match_column', 'match_constraint', 'bool_column', 'parent_children',
     'author_pair', 'age_pair', 'bound_fk', 'unbound_fk', 'want_column',
     'a_relationship', 'BitSelector', 'secret_column', 'username_column',
-    'ascii_column',
+    'ascii_column', 'i4_column',
     # .asyncio
     'SQLAlchemy', 'AsyncSelectPagination', 'async_query', 'SessionWrapper'
 )